Implement library route existence and fix launch bug
Launch bug was due to the config not being managed as part of the state
before the `get_config_from_state` had its handler invoked in the
building process in `main`. To fix it, I moved the config state
management to go before the async process gets launched. This may slow
down launch time on less powerful systems but should't be an issue.